+++ Assets/_core/Script/Live/IDamageable.cs
@@ -2,6 +2,6 @@
 {
     public interface IDamageable
     {
-        void GetHit(float damage);
+        void GetHit(int damage);
     }
 }
